refactor: address PR review feedback for agent instance endpoints
Use withAuthorization in getByKey to correctly scope the authorization check
to the specific entity's processDefinitionId, matching the pattern used by
ElementInstanceServices and other entity services.

Move GET/search tests to AgentInstanceQueryControllerTest following the
*QueryControllerTest convention used across the codebase (e.g., JobQueryControllerTest,
UserTaskQueryControllerTest) where mutation tests live in *ControllerTest and
read/query tests live in *QueryControllerTest.
