+++ zeebe/gateway-rest/src/test/java/io/camunda/zeebe/gateway/rest/controller/AgentInstanceControllerTest.java
@@ -11,27 +11,14 @@ import static org.assertj.core.api.Assertions.assertThat;
 import static org.junit.jupiter.api.Named.named;
 import static org.mockito.ArgumentMatchers.any;
 import static org.mockito.ArgumentMatchers.assertArg;
-import static org.mockito.ArgumentMatchers.eq;
 import static org.mockito.Mockito.verify;
 import static org.mockito.Mockito.verifyNoInteractions;
 import static org.mockito.Mockito.when;
 
-import io.camunda.search.entities.AgentInstanceEntity;
-import io.camunda.search.entities.AgentInstanceEntity.AgentInstanceDefinition;
-import io.camunda.search.entities.AgentInstanceEntity.AgentInstanceLimits;
-import io.camunda.search.entities.AgentInstanceEntity.AgentInstanceMetrics;
-import io.camunda.search.entities.AgentInstanceEntity.AgentInstanceStatus;
-import io.camunda.search.entities.AgentInstanceEntity.AgentInstanceTool;
-import io.camunda.search.exception.CamundaSearchException;
-import io.camunda.search.query.AgentInstanceQuery;
-import io.camunda.search.query.SearchQueryResult;
 import io.camunda.security.api.context.CamundaAuthenticationProvider;
 import io.camunda.service.AgentInstanceServices;
-import io.camunda.service.exception.ErrorMapper;
 import io.camunda.zeebe.gateway.rest.RestControllerTest;
 import io.camunda.zeebe.protocol.impl.record.value.agentinstance.AgentInstanceRecord;
-import java.time.OffsetDateTime;
-import java.util.List;
 import java.util.concurrent.CompletableFuture;
 import java.util.stream.Stream;
 import org.junit.jupiter.api.BeforeEach;
@@ -50,36 +37,6 @@ class AgentInstanceControllerTest extends RestControllerTest {
   private static final String AGENT_INSTANCES_URL = "/v2/agent-instances";
   private static final long ELEMENT_INSTANCE_KEY = 2251799813685248L;
   private static final long AGENT_INSTANCE_KEY = 9007199254741017L;
-  private static final long PROCESS_INSTANCE_KEY = 9007199254741001L;
-  private static final long ROOT_PROCESS_INSTANCE_KEY = 9007199254741000L;
-  private static final long PROCESS_DEFINITION_KEY = 9007199254740992L;
-  private static final OffsetDateTime CREATION_DATE =
-      OffsetDateTime.parse("2024-01-01T10:00:00+00:00");
-  private static final OffsetDateTime LAST_UPDATED_DATE =
-      OffsetDateTime.parse("2024-01-01T10:05:00+00:00");
-  private static final OffsetDateTime COMPLETION_DATE =
-      OffsetDateTime.parse("2024-01-01T10:05:00+00:00");
-
-  private static final AgentInstanceEntity AGENT_INSTANCE_ENTITY =
-      new AgentInstanceEntity(
-          AGENT_INSTANCE_KEY,
-          List.of(ELEMENT_INSTANCE_KEY),
-          AgentInstanceStatus.COMPLETED,
-          new AgentInstanceDefinition("gpt-4o", "openai", "You are a helpful assistant."),
-          new AgentInstanceMetrics(100L, 200L, 3, 5),
-          new AgentInstanceLimits(10000L, 10, 50),
-          List.of(new AgentInstanceTool("search", "Search the web", "searchElement")),
-          "AgentTask",
-          PROCESS_INSTANCE_KEY,
-          ROOT_PROCESS_INSTANCE_KEY,
-          PROCESS_DEFINITION_KEY,
-          "myProcessId",
-          1,
-          "v1",
-          "<default>",
-          CREATION_DATE,
-          LAST_UPDATED_DATE,
-          COMPLETION_DATE);
 
   @MockitoBean private AgentInstanceServices agentInstanceServices;
   @MockitoBean private CamundaAuthenticationProvider authenticationProvider;
@@ -88,184 +45,6 @@ class AgentInstanceControllerTest extends RestControllerTest {
   void setUp() {
     when(authenticationProvider.getCamundaAuthentication())
         .thenReturn(AUTHENTICATION_WITH_DEFAULT_TENANT);
-  }
-
-  @Test
-  void shouldGetAgentInstance() {
-    // given
-    when(agentInstanceServices.getByKey(eq(AGENT_INSTANCE_KEY), any()))
-        .thenReturn(AGENT_INSTANCE_ENTITY);
-
-    // when / then
-    webClient
-        .get()
-        .uri("%s/%d".formatted(AGENT_INSTANCES_URL, AGENT_INSTANCE_KEY))
-        .accept(MediaType.APPLICATION_JSON)
-        .exchange()
-        .expectStatus()
-        .isOk()
-        .expectBody()
-        .json(
-            """
-            {
-              "agentInstanceKey": "%d",
-              "status": "COMPLETED",
-              "definition": {
-                "model": "gpt-4o",
-                "provider": "openai",
-                "systemPrompt": "You are a helpful assistant."
-              },
-              "metrics": {
-                "inputTokens": 100,
-                "outputTokens": 200,
-                "modelCalls": 3,
-                "toolCalls": 5
-              },
-              "limits": {
-                "maxModelCalls": 10,
-                "maxTokens": 10000,
-                "maxToolCalls": 50
-              },
-              "tools": [
-                {
-                  "name": "search",
-                  "description": "Search the web",
-                  "elementId": "searchElement"
-                }
-              ],
-              "elementId": "AgentTask",
-              "processInstanceKey": "%d",
-              "rootProcessInstanceKey": "%d",
-              "processDefinitionKey": "%d",
-              "processDefinitionId": "myProcessId",
-              "processDefinitionVersion": 1,
-              "processDefinitionVersionTag": "v1",
-              "tenantId": "<default>",
-              "creationDate": "2024-01-01T10:00:00.000Z",
-              "lastUpdatedDate": "2024-01-01T10:05:00.000Z",
-              "completionDate": "2024-01-01T10:05:00.000Z",
-              "elementInstanceKeys": ["%d"]
-            }
-            """
-                .formatted(
-                    AGENT_INSTANCE_KEY,
-                    PROCESS_INSTANCE_KEY,
-                    ROOT_PROCESS_INSTANCE_KEY,
-                    PROCESS_DEFINITION_KEY,
-                    ELEMENT_INSTANCE_KEY),
-            JsonCompareMode.LENIENT);
-  }
-
-  @Test
-  void shouldReturnNotFoundForUnknownAgentInstanceKey() {
-    // given
-    final var path = "%s/%d".formatted(AGENT_INSTANCES_URL, AGENT_INSTANCE_KEY);
-    when(agentInstanceServices.getByKey(eq(AGENT_INSTANCE_KEY), any()))
-        .thenThrow(
-            ErrorMapper.mapSearchError(
-                new CamundaSearchException(
-                    "agent instance not found", CamundaSearchException.Reason.NOT_FOUND)));
-
-    // when / then
-    webClient
-        .get()
-        .uri(path)
-        .accept(MediaType.APPLICATION_JSON)
-        .exchange()
-        .expectStatus()
-        .isNotFound()
-        .expectBody()
-        .json(
-            """
-            {
-              "type": "about:blank",
-              "title": "NOT_FOUND",
-              "status": 404,
-              "detail": "agent instance not found",
-              "instance": "%s"
-            }"""
-                .formatted(path),
-            JsonCompareMode.STRICT);
-  }
-
-  @Test
-  void shouldSearchAgentInstancesWithEmptyBody() {
-    // given
-    when(agentInstanceServices.search(any(AgentInstanceQuery.class), any()))
-        .thenReturn(
-            new SearchQueryResult.Builder<AgentInstanceEntity>().total(0).items(List.of()).build());
-
-    // when / then
-    webClient
-        .post()
-        .uri("%s/search".formatted(AGENT_INSTANCES_URL))
-        .accept(MediaType.APPLICATION_JSON)
-        .contentType(MediaType.APPLICATION_JSON)
-        .exchange()
-        .expectStatus()
-        .isOk()
-        .expectBody()
-        .json(
-            """
-            {
-              "items": [],
-              "page": { "totalItems": 0 }
-            }
-            """,
-            JsonCompareMode.LENIENT);
-  }
-
-  @Test
-  void shouldSearchAgentInstances() {
-    // given
-    when(agentInstanceServices.search(any(AgentInstanceQuery.class), any()))
-        .thenReturn(
-            new SearchQueryResult.Builder<AgentInstanceEntity>()
-                .total(1)
-                .startCursor("f")
-                .endCursor("v")
-                .items(List.of(AGENT_INSTANCE_ENTITY))
-                .build());
-
-    // when / then
-    webClient
-        .post()
-        .uri("%s/search".formatted(AGENT_INSTANCES_URL))
-        .accept(MediaType.APPLICATION_JSON)
-        .contentType(MediaType.APPLICATION_JSON)
-        .bodyValue(
-            """
-            {
-              "filter": {
-                "status": { "$eq": "COMPLETED" }
-              }
-            }
-            """)
-        .exchange()
-        .expectStatus()
-        .isOk()
-        .expectBody()
-        .json(
-            """
-            {
-              "items": [
-                {
-                  "agentInstanceKey": "%d",
-                  "status": "COMPLETED",
-                  "processDefinitionId": "myProcessId",
-                  "processDefinitionVersion": 1,
-                  "processDefinitionVersionTag": "v1"
-                }
-              ],
-              "page": {
-                "totalItems": 1,
-                "startCursor": "f",
-                "endCursor": "v"
-              }
-            }
-            """
-                .formatted(AGENT_INSTANCE_KEY),
-            JsonCompareMode.LENIENT);
   }
 
   @Test
